# packages/api-client/openapi/v1.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Shellwright control plane",
    "description": "Organisations, apps, and immutable configuration versions.\n\nErrors are RFC 9457 problem documents. The `type` URI and the `code` extension are stable; the `title` and `detail` are for people and may be reworded.",
    "version": "v1"
  },
  "paths": {
...
    "/v1/orgs": {
      "get": {
        "tags": ["Organisations"],
        "summary": "List the organisations the caller belongs to.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrgResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Organisations"],
        "summary": "Create an organisation and become its owner.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrgRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrgResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/orgs/{orgId}/workspaces": {
      "get": {
        "tags": ["Organisations"],
        "summary": "List an organisation's workspaces.",
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkspaceResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Organisations"],
        "summary": "Create a workspace.",
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkspaceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/orgs/{orgId}/members": {
      "get": {
        "tags": ["Organisations"],
        "summary": "List an organisation's members.",
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MemberResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/orgs/{orgId}/members/{userId}": {
      "put": {
        "tags": ["Organisations"],
        "summary": "Grant or change a member's role.",
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetMemberRoleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateOrgRequest": {
        "required": ["name", "slug"],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Display name."
          },
          "slug": {
            "type": "string",
            "description": "URL-safe identifier, generated from the name when omitted.",
            "nullable": true
          }
        },
        "description": "Create an organisation."
      },
      "CreateWorkspaceRequest": {
        "required": ["name", "slug"],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Display name."
          },
          "slug": {
            "type": "string",
            "description": "URL-safe identifier, generated from the name when omitted.",
            "nullable": true
          }
        },
        "description": "Create a workspace."
      },
...
      "MemberResponse": {
        "required": ["userId", "email", "role"],
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "The member.",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "description": "Their address."
          },
          "role": {
            "type": "string",
            "description": "Their role."
          }
        },
        "description": "A membership as the API reports it."
      },
      "OrgResponse": {
        "required": ["id", "name", "slug", "plan", "role"],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Identifier.",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "description": "Display name."
          },
          "slug": {
            "type": "string",
            "description": "URL-safe identifier."
          },
          "plan": {
            "type": "string",
            "description": "Billing plan."
          },
          "role": {
            "type": "string",
            "description": "The caller's effective role."
          }
        },
        "description": "An organisation as the API reports it."
      },
      "OrgRole": {
        "enum": ["Viewer", "Developer", "Admin", "Owner"],
        "description": "Membership roles, ordered from least to most privileged.",
        "default": "Developer"
      },
...
      "SetMemberRoleRequest": {
        "required": ["role"],
        "type": "object",
        "properties": {
          "role": {
            "$ref": "#/components/schemas/OrgRole"
          }
        },
        "description": "Change somebody's role."
      },
...
      "WorkspaceResponse": {
        "required": ["id", "orgId", "name", "slug"],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Identifier.",
            "format": "uuid"
          },
          "orgId": {
            "type": "string",
            "description": "Owning organisation.",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "description": "Display name."
          },
          "slug": {
            "type": "string",
            "description": "URL-safe identifier."
          }
        },
        "description": "A workspace as the API reports it."
      }
    }
  },
...
}